requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
pypdf
openai